Create score tracker, winner and round message
Create divs in index.html for displaying who won the round, what
the current score is, and when someone reaches 5, who that person was.
In javascript.js, add paragraph element to each of these divs, then
add text content to each of the divs based on the outcome. then
append the paragraph element to the respective div. RoundReport gets
added inside conditional. Score and conditional for final winner is
added outside the conditionals.